+++ manual_test.sh
@@ -1,0 +1,10 @@
+echo "== TESTING -f test.json ==" && \
+cargo run -- -f test.json && \
+echo "== TESTING -c test.toml ==" && \
+cargo run -- -c test.toml && \
+echo "== TESTING -c . ==" && \
+cargo run -- -c . && \
+echo ""
+echo "== ======= =="
+echo "== SUCCESS =="
+echo "== ======= =="