+++ src/request.rs
@@ -1,10 +1,24 @@
 use reqwest::blocking::Client;
 use reqwest::header::AUTHORIZATION;
 use reqwest::header::CONTENT_TYPE;
+use reqwest::header::USER_AGENT;
+use serde::Deserialize;
 use serde_json::json;
 use std::collections::HashMap;
 
-const URL: &str = "https://api.linear.app/graphql";
+const LINEAR_URL: &str = "https://api.linear.app/graphql";
+const CARGO_URL: &str = "https://crates.io/api";
+const VERSIONS_URL: &str = "/v1/crates/linear_templater/versions";
+
+#[derive(Deserialize)]
+struct CargoResponse {
+    versions: Vec<Version>,
+}
+
+#[derive(Deserialize)]
+struct Version {
+    num: String,
+}
 
 pub fn gql(
     token: String,
@@ -16,7 +30,7 @@ pub fn gql(
     let body = json!({"query": query, "variables": variables});
 
     let response = Client::new()
-        .post(URL)
+        .post(LINEAR_URL)
         .header(CONTENT_TYPE, "application/json")
         .header(AUTHORIZATION, authorization)
         .json(&body)
@@ -25,6 +39,26 @@ pub fn gql(
 
     if response.status().is_success() {
         Ok(response.text().or(Err("Could not read response text"))?)
+    } else {
+        Err(format!("Error: {:#?}", response.text()))
+    }
+}
+
+/// Get latest version number from Cargo.io
+pub fn get_latest_version() -> Result<String, String> {
+    let request_url = format!("{CARGO_URL}{VERSIONS_URL}");
+
+    let response = Client::new()
+        .get(request_url)
+        .header(USER_AGENT, "GPTO")
+        .send()
+        .or(Err("Did not get response from server"))?;
+
+    if response.status().is_success() {
+        let cr: CargoResponse =
+            serde_json::from_str(&response.text().or(Err("Could not read response text"))?)
+                .or(Err("Could not serialize to CargoResponse"))?;
+        Ok(cr.versions.first().unwrap().num.clone())
     } else {
         Err(format!("Error: {:#?}", response.text()))
     }