+++ CHANGELOG.md
@@ -2,11 +2,15 @@
 
 ## Unreleased
 
+## 2023-03-05 v0.1.1
+
 - Refactored things and stuff
 - Added CI and badges
 - Added README and this CHANGELOG
 - Changed from create_tickets to create_issues
 - Added project_id to TOML and JSON
+- Can create tickets from all TOML files in a directory
+- Check for the latest version on crates.io
 
 ## 2023-03-04 v0.1.0